feat: ajout de la classe SearchBox
