+++ css/main/recipes.css
@@ -1,13 +1,11 @@
-#recipes>article>div.recipe-title>span>i {
-    margin-right: .3rem;
-}
-
 #recipes>article {
     background-color: #E7E7E7;
     border-radius: .3125rem;
-    height: 23.75rem;
+    height: auto;
     margin-bottom: 2.5rem;
+    overflow: hidden;
     position: relative;
+    text-overflow: ellipsis;
     width: 23.75rem;
 }
 
@@ -29,6 +27,10 @@
     font-weight: 700;
 }
 
+#recipes>article>div.recipe-title>span>i {
+    margin-right: .3rem;
+}
+
 #recipes>article>div.recipes-info {
     display: flex;
     margin-left: 1rem;
@@ -40,18 +42,19 @@
     width: 50%;
 }
 
-#recipes>article>div.recipes-info>div.recipes-ingredients>p {
-    font-size: .8rem;
-    margin: .125rem 0;
-}
-
+#recipes>article>div.recipes-info>div.recipes-ingredients,
 #recipes>article>div.recipes-info>div.recipes-instructions>span {
     -webkit-box-orient: vertical;
-    -webkit-line-clamp: 8;
+    -webkit-line-clamp: 7;
     display: -webkit-box;
     font-size: .8rem;
     overflow: hidden;
     text-overflow: ellipsis;
+}
+
+#recipes>article>div.recipes-info>div.recipes-ingredients>p {
+    font-size: .8rem;
+    margin: .125rem 0;
 }
 
 #recipes>article>img {
